Add weak statements
Weak statements do not constrain the selection of
their dependencies. This is useful for empty
(unit) statements that do not match any nodes. In
this case we do not want to prevent the whole
statement to match no node.

For example, in the following code:

```
    "main" {{}}
```

If "main" does not have grandchildren, we still
want the query to match "main" and "main"'s
children.